requirements.txt
requests
pytest
PyYAML
